docker-compose/sqlite_to_postgres/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
psycopg==3.1.18
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
typing-extensions==4.12.2
    # via psycopg
